+++ src/detection/locale/locale.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#ifndef FF_INCLUDED_detection_locale_locale
+#define FF_INCLUDED_detection_locale_locale
+
+#include "fastfetch.h"
+
+void ffDetectLocale(FFstrbuf* result);
+
+#endif
